Leave the build with nothing to say
Twenty-four warnings stood between a new one and being noticed. None of
them was a fault, but each had to be read to find that out, and that was
the cost they were quietly charging.

Seventeen were about documentation comments: members without one, two
comments left standing above methods that are themselves commented out,
and two references naming something the compiler cannot resolve. The two
that documented commented-out code are ordinary comments now, since what
they explain is not there to be documented.

The seven about code were each safe, for reasons worth writing down:

- The key for a new check ORs a random lower half into a rising count.
  The random part can never be negative, but it was widened as a signed
  number, and a negative one would have filled the entire upper half with
  ones and buried the count. It goes through uint now, so the bits are
  laid in as they are.
- A registry lookup already promises with NotNullWhen that it fills in its
  result when it succeeds. The variable it filled was declared as never
  null, so the promise was made in two places and kept in one. The
  variable is nullable now and the promise does the work.
- The three things the boss rooms are handed read the two-player save,
  which is made after them, because that save asks them in turn for the
  room a player waits in. Neither can be first; both exist long before
  any of the three is called.
- A tuning value for interpolation was set and read by nothing at all.

Nothing about what the mod does changes.

+++ SSMP/Game/Client/Save/CoopSave.Check.cs
@@ -233,7 +233,9 @@ internal partial class CoopSave {
     /// too.
     /// </summary>
     private ulong NewCheckKey() {
-        var key = (((_highestCheckKey >> 16) + 1) << 16) | (ulong) UnityEngine.Random.Range(1, 0x10000);
+        // Through uint, so that the lower bits are laid in as they are. Random.Range never returns a negative here,
+        // but a signed value widened to ulong would fill the whole upper half with ones and bury the count above
+        var key = (((_highestCheckKey >> 16) + 1) << 16) | (uint) UnityEngine.Random.Range(1, 0x10000);
         _highestCheckKey = key;
         return key;
     }
